Updated classified post/edit & finish

+++ server/controllers/api/classified/put.coffee
@@ -5,6 +5,10 @@ module.exports = (request, response, next) ->
   File = global.helpers.file
   response.contentType 'application/json'
 
+  id = request.params.id
+  authHash = request.query.authHash
+
+  console.log id, authHash
   captchaFail = ->
     response.status 401
     response.end '"captcha failed"'
@@ -16,54 +20,54 @@ module.exports = (request, response, next) ->
     form.multiples = true
     form.maxFieldsSize = 10 * 1024 * 1024 # 2MB
 
-    # Setup error handler. This function gets called whenever there is an
-    # error while processing the form.
-    form.on 'error', (error) ->
-      response.status 400
-      response.end JSON.stringify error
+    Classified.get id, (error, classified) ->
+      if classified.owner is not request.user._id
+        response.status 401
+        return response.end '"not owner"'
+      if classified.guest and classified.authHash is not authHash
+        response.status 401
+        return response.end '"not owner"'
 
-    # Start parsing the form
-    form.parse request, (error, fields, filesRequest) ->
-      if error then return next error
+      # Setup error handler. This function gets called whenever there is an
+      # error while processing the form.
+      form.on 'error', (error) ->
+        response.status 400
+        response.end JSON.stringify error
 
-      data = JSON.parse fields.data
+      # Start parsing the form
+      form.parse request, (error, fields, filesRequest) ->
+        if error then return next error
 
-      files = filesRequest['files[]']
-      filesToDelete = data.filesToDelete or []
-      images = data.images
-      File.delete data.filesToDelete
-      images.splice (images.indexOf file), 1 for file in filesToDelete
-      data.images = images
+        data = JSON.parse fields.data
 
-      File.upload files, (error, files=[]) ->
-
-        if error
-          response.status 400
-          return response.end JSON.stringify error
-
-        images = data.images or []
-        images.push file for file in files
+        files = filesRequest['files[]']
+        filesToDelete = data.filesToDelete or []
+        images = data.images
+        File.delete data.filesToDelete
+        images.splice (images.indexOf file), 1 for file in filesToDelete
         data.images = images
 
-        Classified.update data, request.user, (error, classified) ->
-          # If error was set, then nothing was updated.
-          # Send a 400 Bad Request to the client
+        File.upload files, (error, files=[]) ->
+
           if error
-            File.delete files
             response.status 400
             return response.end JSON.stringify error
 
-          # # Send classified-online email
-          # toAddress = (classified.contact or {}).email || (request.user or {}).email
-          # Email = global.controllers.helpers.email
-          # Email.sendTemplate toAddress, 'classified-online',
-          #   subject: "Your classified is now online!"
-          #   classified: classified
+          images = data.images or []
+          images.push file for file in files
+          data.images = images
 
-          # If a classified was updated, then return it to the client.
-          # The returned classified will contain the id parameter which
-          # gets set by the database
-          if classified then return response.end JSON.stringify classified
+          Classified.update data, request.user, (error, classified) ->
+            # If error was set, then nothing was updated.
+            if error
+              File.delete files
+              response.status 400
+              return response.end JSON.stringify error
+
+            # If a classified was updated, then return it to the client.
+            # The returned classified will contain the id parameter which
+            # gets set by the database
+            if classified then return response.end JSON.stringify classified
 
   reCaptcha = global.helpers.reCaptcha
   reCaptcha.verify request, captchaSuccess, captchaFail
